forme now has distinct types of page. PageLocation and PageContainer. a page location is just a path to a page that is responsible for handling a form. This allows forms to be spread across different scripts. A PageContainer is for designing your entire form within 1 script. Pages can be added statically or dynamically via a callback. Dynamic pages only get built when visiting a speciffic page.
Added ability to reset the form data

forme now handles validation of the form journey. For example visiting page 3 before you have got their naturally, will reset the form.

values are now stored per page

added store options to _fetchValues and _findValue. This gives the option to build/find from the store. This means we can produce a solid object of the entire forms data if we need.

added "skip" processing to all routines. This is request._reset. this allows us to cancel the current form, without having to do gnarly promises.

ton of helper utilities for all manor of things

added "input.submit()" which allows us to to define a submit button. This does not have a special action (yet) but could be designed to fast foward form submit if we wanted.

added a ton of throw werrors when trying to use some apis that are incompat.

result.destination now removed uneeded query params