split handlers into process and validate categories. validate are async and process are synchronous. This is so that we can run all process handlers on a value for a given input and return. For example if we want to fetch the value early, before the form has finished building, we need a way to run various handlers. A prime example would be converting the value to int() float() etc. Currently there is no way to add custom process handlers, and it is limited to the Convert handler.
refactored how the form sets and gets values on an input. The functions now expect just the input name. This is so we can deal with inputs regardless of where they are in the build/validate/eetc pipeline.

forme now runs "execution" handlers instead of "validate" handlers. Execution handlers are an ordered of list of both validate and process handlers. The order that we define our input api's still matters, so we preserve this as a combined list.

fixed issue where input state value was not updating all the time after validate handlers!

added method for *manual* processing of value to input.

can now use form.value() during building and the value will be as exepected (e.g. int(), float()). Still can only call form.value() for inputs that exist in the form. Could be expanded later to get an input value that hasnt been added yet, but was submitted already from a previous form load.

added *strict* flag to various input api's. Defaults to false. This indicates if comparison on values should be done with == or ===.